Added navigation from cities page
